Changed the contact info to use actions instead of API calls, as well as moving the page into a component
